+++ src/Novell.Directory.Ldap.NETStandard/Connection.cs
@@ -31,6 +31,7 @@
 //
 
 using System;
+using System.Collections;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
@@ -1122,11 +1123,12 @@ namespace Novell.Directory.Ldap
                 // Dispose will break forcefully the Read.
                 // We could use a ReadTimeout for socket - but this will only make stopping the thread take longer
                 // And we don't care if we just kill the socket stream as we don't plan to reuse the stream after stop
-                // the stream Dispose used to be called from Connection dispose but only when a Bind is succesful which was causing
-                // the Dispose to hang un unsuccesful bind
+                // the stream Dispose used to be called from Connection dispose but only when a Bind is successful which was causing
+                // the Dispose to hang un unsuccessful bind
                 // So, yeah isStopping flag is pretty much useless as there are very small chances that it will be hit
                 var socketStream = _enclosingInstance._inStream;
                 socketStream?.Dispose();
+
                 _enclosedThread.Join();
             }