+++ src/Novell.Directory.Ldap.NETStandard/SearchExtensions/VirtualListViewControlSearchExtensions.cs
@@ -1,0 +1,46 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using JetBrains.Annotations;
+using Novell.Directory.Ldap.Controls;
+
+namespace Novell.Directory.Ldap.SearchExtensions
+{
+    /// <summary>
+    /// Extensions methods to <see cref="ILdapConnection"/> to be able to do searches using <see cref="LdapVirtualListControl"/>
+    /// using <see cref="VirtualListViewControlHandler"/>.
+    /// </summary>
+    public static class VirtualListViewControlSearchExtensions
+    {
+        public static List<LdapEntry> SearchUsingVlv(
+            [NotNull] this ILdapConnection ldapConnection,
+            [NotNull] LdapSortControl sortControl,
+            [NotNull] SearchOptions options, 
+            int pageSize)
+        {
+            if (ldapConnection == null) throw new ArgumentNullException(nameof(ldapConnection));
+            if (sortControl == null) throw new ArgumentNullException(nameof(sortControl));
+            if (options == null) throw new ArgumentNullException(nameof(options));
+            if (pageSize <= 0) throw new ArgumentOutOfRangeException(nameof(pageSize));
+
+            return new VirtualListViewControlHandler(ldapConnection)
+                .SearchUsingVlv(sortControl, options, pageSize);
+        }
+
+        public static List<T> SearchUsingVlv<T>(
+            [NotNull] this ILdapConnection ldapConnection,
+            [NotNull] LdapSortControl sortControl,
+            [NotNull] Func<LdapEntry, T> converter,
+            [NotNull] SearchOptions options,
+            int pageSize)
+        {
+            if (ldapConnection == null) throw new ArgumentNullException(nameof(ldapConnection));
+            if (sortControl == null) throw new ArgumentNullException(nameof(sortControl));
+            if (options == null) throw new ArgumentNullException(nameof(options));
+            if (pageSize <= 0) throw new ArgumentOutOfRangeException(nameof(pageSize));
+
+            return new VirtualListViewControlHandler(ldapConnection)
+                .SearchUsingVlv<T>(sortControl, converter, options, pageSize);
+        }
+    }
+}